thread: fix archival/deletion event
